[svn] ConEmu sources
git-svn-id: https://conemu-maximus5.googlecode.com/svn@1240 92116c04-ab5e-ca4c-3a8b-6f4eadf4b4a1

+++ src/ConEmu/Menu.cpp
@@ -455,7 +455,7 @@ void CConEmuMenu::OnNewConPopupMenuRClick(HMENU hMenu, UINT nItemPos)
 
 	DestroyMenu(hPopup);
 
-	if (nRetID >= (nStartID+1) && nRetID <= (int)(nStartID+nLastID))
+	if (nRetID >= (int)(nStartID+1) && nRetID <= (int)(nStartID+nLastID))
 	{
 		// Need to close parentmenu
 		if (hMenuWnd)
